requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Django==4.1.7
django-celery-beat==2.4.0
django-timezone-field==5.0
future==0.18.3
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-telegram-bot==12.0.0b1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
